+++ apps/io-wallet-user-func/src/infra/azure/functions/add-wallet-instance-to-validation-queue.ts
@@ -1,0 +1,6 @@
+import { AddWalletInstanceToValidationQueueHandler } from "@/infra/handlers/add-wallet-instance-to-validation-queue";
+import { azureFunction } from "@pagopa/handler-kit-azure-func";
+
+export const AddWalletInstanceToValidationQueueFunction = azureFunction(
+  AddWalletInstanceToValidationQueueHandler,
+);