+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/get-current-wallet-instance.status.spec.ts
@@ -20,7 +20,6 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
   const walletInstanceRepository: WalletInstanceRepository = {
     batchPatch: () => TE.left(new Error("not implemented")),
     get: () => TE.left(new Error("not implemented")),
-    getAllByUserId: () => TE.left(new Error("not implemented")),
     getLastByUserId: () =>
       TE.right(
         O.some({
@@ -37,15 +36,16 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
           userId: "AAA" as FiscalCode,
         }),
       ),
+    getValidByUserIdExcludingOne: () => TE.left(new Error("not implemented")),
     insert: () => TE.left(new Error("not implemented")),
   };
 
   const req = {
     ...H.request("https://wallet-provider.example.org"),
-    body: {
-      fiscal_code: "GSPMTA98L25E625O",
+    headers: {
+      "fiscal-code": "GSPMTA98L25E625O",
     },
-    method: "POST",
+    method: "GET",
   };
 
   const telemetryClient: appInsights.TelemetryClient = {
@@ -76,14 +76,43 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
     });
   });
 
-  it("should return a 422 HTTP response on invalid body", async () => {
+  it("should return a 400 HTTP response when fiscal_code header is missing", async () => {
     const req = {
       ...H.request("https://wallet-provider.example.org"),
-      body: {
-        foo: "GSPMTA98L25E625O",
+      headers: {
+        fiscalcode: "GSPMTA98L25E625O",
       },
-      method: "POST",
+      method: "GET",
     };
+
+    const handler = GetCurrentWalletInstanceStatusHandler({
+      input: req,
+      inputDecoder: H.HttpRequest,
+      logger,
+      telemetryClient,
+      walletInstanceRepository,
+    });
+
+    await expect(handler()).resolves.toEqual({
+      _tag: "Right",
+      right: expect.objectContaining({
+        headers: expect.objectContaining({
+          "Content-Type": "application/problem+json",
+        }),
+        statusCode: 400,
+      }),
+    });
+  });
+
+  it("should return a 422 HTTP response when fiscal_code header is not a valid fiscal code", async () => {
+    const req = {
+      ...H.request("https://wallet-provider.example.org"),
+      headers: {
+        "fiscal-code": "foo",
+      },
+      method: "GET",
+    };
+
     const handler = GetCurrentWalletInstanceStatusHandler({
       input: req,
       inputDecoder: H.HttpRequest,
@@ -107,8 +136,8 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
     const walletInstanceRepository: WalletInstanceRepository = {
       batchPatch: () => TE.left(new Error("not implemented")),
       get: () => TE.left(new Error("not implemented")),
-      getAllByUserId: () => TE.left(new Error("not implemented")),
       getLastByUserId: () => TE.right(O.none),
+      getValidByUserIdExcludingOne: () => TE.left(new Error("not implemented")),
       insert: () => TE.left(new Error("not implemented")),
     };
     const handler = GetCurrentWalletInstanceStatusHandler({
@@ -135,8 +164,9 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
       {
         batchPatch: () => TE.left(new Error("not implemented")),
         get: () => TE.left(new Error("not implemented")),
-        getAllByUserId: () => TE.left(new Error("not implemented")),
         getLastByUserId: () => TE.left(new Error("failed on getLastByUserId!")),
+        getValidByUserIdExcludingOne: () =>
+          TE.left(new Error("not implemented")),
         insert: () => TE.left(new Error("not implemented")),
       };
     const handler = GetCurrentWalletInstanceStatusHandler({
@@ -164,8 +194,9 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
       {
         batchPatch: () => TE.left(new Error("not implemented")),
         get: () => TE.left(new Error("not implemented")),
-        getAllByUserId: () => TE.left(new Error("not implemented")),
         getLastByUserId: () => TE.left(new ServiceUnavailableError("foo")),
+        getValidByUserIdExcludingOne: () =>
+          TE.left(new Error("not implemented")),
         insert: () => TE.left(new Error("not implemented")),
       };
     const handler = GetCurrentWalletInstanceStatusHandler({