+++ apps/io-wallet-user-func/src/wallet-instance.ts
@@ -1,3 +1,4 @@
+import * as E from "fp-ts/Either";
 import * as O from "fp-ts/Option";
 import * as RTE from "fp-ts/ReaderTaskEither";
 import * as RA from "fp-ts/ReadonlyArray";
@@ -5,8 +6,10 @@ import * as TE from "fp-ts/TaskEither";
 import { flow, pipe } from "fp-ts/function";
 import { EntityNotFoundError } from "io-wallet-common/error";
 import {
+  RevocationReason,
   WalletInstance,
   WalletInstanceValid,
+  WalletInstanceValidWithAndroidCertificatesChain,
 } from "io-wallet-common/wallet-instance";
 
 class RevokedWalletInstance extends Error {
@@ -32,12 +35,13 @@ export interface WalletInstanceRepository {
     id: WalletInstance["id"],
     userId: WalletInstance["userId"],
   ) => TE.TaskEither<Error, O.Option<WalletInstance>>;
-  getAllByUserId: (
-    userId: WalletInstance["userId"],
-  ) => TE.TaskEither<Error, O.Option<WalletInstance[]>>;
   getLastByUserId: (
     userId: WalletInstance["userId"],
   ) => TE.TaskEither<Error, O.Option<WalletInstance>>;
+  getValidByUserIdExcludingOne: (
+    walletInstanceId: WalletInstance["id"],
+    userId: WalletInstance["userId"],
+  ) => TE.TaskEither<Error, O.Option<WalletInstanceValid[]>>;
   insert: (walletInstance: WalletInstanceValid) => TE.TaskEither<Error, void>;
 }
 
@@ -90,11 +94,35 @@ export const getValidWalletInstance: (
       ),
     );
 
+export const getValidWalletInstanceWithAndroidCertificatesChain: (
+  id: WalletInstance["id"],
+  userId: WalletInstance["userId"],
+) => RTE.ReaderTaskEither<
+  WalletInstanceEnvironment,
+  Error,
+  WalletInstanceValidWithAndroidCertificatesChain
+> = (id, userId) =>
+  pipe(
+    getValidWalletInstance(id, userId),
+    RTE.chainEitherK(
+      flow(
+        WalletInstanceValidWithAndroidCertificatesChain.decode,
+        E.mapLeft(
+          () =>
+            new Error(
+              "Wallet Instance does not have a certificate chain for android",
+            ),
+        ),
+      ),
+    ),
+  );
+
 export const revokeUserWalletInstances: (
   userId: WalletInstance["userId"],
   walletInstancesId: readonly WalletInstance["id"][],
+  reason: RevocationReason,
 ) => RTE.ReaderTaskEither<WalletInstanceEnvironment, Error, void> =
-  (userId, walletInstancesId) =>
+  (userId, walletInstancesId, reason) =>
   ({ walletInstanceRepository }) =>
     walletInstancesId.length
       ? walletInstanceRepository.batchPatch(
@@ -110,6 +138,11 @@ export const revokeUserWalletInstances: (
                 op: "add",
                 path: "/revokedAt",
                 value: new Date(),
+              },
+              {
+                op: "add",
+                path: "/revocationReason",
+                value: reason,
               },
             ],
           })),
@@ -128,18 +161,14 @@ const getUserValidWalletInstancesIdExceptOne: (
   (userId, walletInstanceId) =>
   ({ walletInstanceRepository }) =>
     pipe(
-      walletInstanceRepository.getAllByUserId(userId),
+      walletInstanceRepository.getValidByUserIdExcludingOne(
+        walletInstanceId,
+        userId,
+      ),
       TE.map(
         O.fold(
           () => [],
-          flow(
-            RA.filterMap((walletInstance) =>
-              walletInstance.id !== walletInstanceId &&
-              walletInstance.isRevoked === false
-                ? O.some(walletInstance.id)
-                : O.none,
-            ),
-          ),
+          flow(RA.filterMap((walletInstance) => O.some(walletInstance.id))),
         ),
       ),
     );
@@ -147,13 +176,30 @@ const getUserValidWalletInstancesIdExceptOne: (
 export const revokeUserValidWalletInstancesExceptOne: (
   userId: WalletInstance["userId"],
   walletInstanceId: WalletInstance["id"],
+  reason: RevocationReason,
 ) => RTE.ReaderTaskEither<WalletInstanceEnvironment, Error, void> = (
   userId,
   walletInstanceId,
+  reason,
 ) =>
   pipe(
     getUserValidWalletInstancesIdExceptOne(userId, walletInstanceId),
     RTE.chain((validWalletInstances) =>
-      revokeUserWalletInstances(userId, validWalletInstances),
+      revokeUserWalletInstances(userId, validWalletInstances, reason),
     ),
+  );
+
+export const filterValidWithAndroidCertificatesChain = (
+  walletInstances: WalletInstance[],
+): RTE.ReaderTaskEither<
+  void,
+  Error,
+  readonly WalletInstanceValidWithAndroidCertificatesChain[]
+> =>
+  pipe(
+    walletInstances,
+    RA.filterMap(
+      flow(WalletInstanceValidWithAndroidCertificatesChain.decode, O.getRight),
+    ),
+    RTE.of,
   );