+++ packages/io-wallet-common/src/wallet-instance.ts
@@ -2,7 +2,7 @@ import { IsoDateFromString } from "@pagopa/ts-commons/lib/dates";
 import { FiscalCode, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import * as t from "io-ts";
 
-import { DeviceDetails } from "./device-details";
+import { AndroidDeviceDetails, DeviceDetails } from "./device-details";
 import { JwkPublicKey } from "./jwk";
 
 const WalletInstanceBase = t.intersection([
@@ -18,7 +18,7 @@ const WalletInstanceBase = t.intersection([
   }),
 ]);
 
-const WalletInstanceValid = t.intersection([
+export const WalletInstanceValid = t.intersection([
   WalletInstanceBase,
   t.type({
     isRevoked: t.literal(false),
@@ -27,11 +27,21 @@ const WalletInstanceValid = t.intersection([
 
 export type WalletInstanceValid = t.TypeOf<typeof WalletInstanceValid>;
 
+export const RevocationReason = t.union([
+  t.literal("CERTIFICATE_REVOKED_BY_ISSUER"),
+  t.literal("NEW_WALLET_INSTANCE_CREATED"),
+  t.literal("REVOKED_BY_USER"),
+]);
+export type RevocationReason = t.TypeOf<typeof RevocationReason>;
+
 const WalletInstanceRevoked = t.intersection([
   WalletInstanceBase,
   t.type({
     isRevoked: t.literal(true),
     revokedAt: IsoDateFromString,
+  }),
+  t.partial({
+    revocationReason: RevocationReason,
   }),
 ]);
 
@@ -41,3 +51,19 @@ export const WalletInstance = t.union([
 ]);
 
 export type WalletInstance = t.TypeOf<typeof WalletInstance>;
+
+export const WalletInstanceValidWithAndroidCertificatesChain = t.intersection([
+  WalletInstanceValid,
+  t.type({
+    deviceDetails: t.intersection([
+      AndroidDeviceDetails,
+      t.type({
+        x509Chain: t.array(t.string),
+      }),
+    ]),
+  }),
+]);
+
+export type WalletInstanceValidWithAndroidCertificatesChain = t.TypeOf<
+  typeof WalletInstanceValidWithAndroidCertificatesChain
+>;