+++ apps/io-wallet-support-func/CHANGELOG.md
@@ -1,5 +1,25 @@
 # io-wallet-support-func
 
+## 1.3.4
+
+### Patch Changes
+
+- d829f56: Add revocation_reason field to Wallet Instance
+
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies [07d296c]
+  - io-wallet-common@1.2.1
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - io-wallet-common@1.1.2
+
 ## 1.3.1
 
 ### Patch Changes